Revert "[Feat] 숙소 상세 페이지 및 테스트 컨테이너 적용 (#10)"
This reverts commit c45f8ac93243031f06024d84fbbf9b5438996c2e.

+++ backend/src/main/java/com/team3/airdnd/payment/domain/Payment.java
@@ -1,6 +1,6 @@
 package com.team3.airdnd.payment.domain;
 
-import com.team3.airdnd.accommodation.domain.Reservation;
+import com.team3.airdnd.accomodation.domain.Reservation;
 import jakarta.persistence.*;
 import lombok.*;
 
@@ -12,7 +12,6 @@ import java.time.LocalDateTime;
 @AllArgsConstructor
 @NoArgsConstructor
 public class Payment {
-
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
@@ -25,7 +24,7 @@ public class Payment {
     @JoinColumn(name = "payment_method_id", nullable = false)
     private PaymentMethod paymentMethod;
 
-    private Long amount;
+    private Integer amount;
 
     @Column(name = "paid_at", nullable = false)
     private LocalDateTime paidAt;
